Update Discipline.cs
fixed

+++ src/UMS.Domain/Entities/Education/Discipline.cs
@@ -3,8 +3,8 @@
     public class Discipline : Auditable
     {
         public string Name { get; set; }
-        public int DepartmentId { get; set; }
-        public int TeacherId { get; set; }
+        public long DepartmentId { get; set; }
+        public long TeacherId { get; set; }
         public int LectureHours { get; set; }
         public int PracticeHours { get; set; }
     }
